requirements.txt
librosa==0.9.2
numpy==1.24.3
pandas==1.3.5
torchaudio

pyannote.audio
pyannote.core==4.5
pyannote.database==4.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==2.3
PyAudio==0.2.13
pydub==0.25.1
SpeechRecognition==3.10.0
